Add support for python 3.11+ (#5)
The cython generated C code is different for 3.11 vs 3.10, so
relying on pip using an sdist with that file included doesn't work.
That is to say, that "pip install sha256" doesn't work on 3.11 with
the 0.3 release on pypi.

This change updates setup.py to the current recommendation from
cython (which removes the sdist step).  It adds a pyproject.toml
so that cython is recognized as a build-time dependency in recent
pip versions.  Package classifiers are updated to include 3.11
and 3.12, and dropping the EOL 3.7. Finally the sha256 package
version is updated to 1.0 to prepare for a future release.